pyproject.toml
[tool.poetry]
name = "anime-pgen"
version = "1.0.1"
description = "cli-утилита для генерации превью-изображений по данным Шикимори"
authors = ["Vladimir Levin <opaozhub@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-only"
repository = "https://github.com/shikimori/preview-generator"
documentation = "https://github.com/shikimori/preview-generator/blob/main/DOCS.md"

[tool.poetry.scripts]
pgen = "anime_pgen.main:app"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.0.1"
typer = { extras = ["all"], version = "^0.4.0" }
click = "8.1.3"
requests = "^2.27.1"
ratelimiter = "^1.2.0"
colour = "^0.1.5"
numpy = "^1.22.3"
PyYAML = "^6.0"
joblib = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
